feat: add fully functioning sqrt method and error handling for it
It uses BigInt on the inside, and warns the user if a number was passed,
because it automatically converts it to the scientific notation, where
precision gets less.
The current sqrt implementation is also not 100% perfect, as the last
couple digits are almost certainly incorrect, which happens due to
irrational numbers not having the exact sqrt value. Thus - the precision
is set huge (17 digits as of now).